Modify TCP Client to return an error instead of blocking when the network Send-Q is full
When the physical network connection is lost (without closing the socket) the messages wait in the operating send queue. When this queue is full the write operation will block, preventing the application from responding to the loss of network connection. This change allows the write operation to return an error, although it will still take time for the send queue to become full.